chore: don't externalize vscode and move util to utils vscode
