+++ packages/salesforcedx-utils/src/helpers/index.ts
@@ -5,5 +5,4 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 export { extractJsonObject } from './extractJsonObject';
-export { ensureCurrentWorkingDirIsProjectPath } from './ensureCurrentWorkingDirIsProjectPath';
 export { stripAnsiInJson } from './utils';
