+++ packages/salesforcedx-vscode-apex-replay-debugger/src/adapter/debugConfigurationProvider.ts
@@ -52,7 +52,7 @@ export class DebugConfigurationProvider implements vscode.DebugConfigurationProv
     }
 
     if (vscode.workspace && vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0]) {
-      config.projectPath = vscode.workspace.workspaceFolders[0].uri.toString();
+      config.projectPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
     }
 
     if (this.salesforceApexExtension && this.salesforceApexExtension.exports) {
